requirements.txt
pycryptodome
opencv-python
numpy
